Add Life Create
入力したライフを反映できるようにしました
